+++ tests/test_util.py
@@ -5,9 +5,8 @@ import os
 import sys
 import shelve
 
-from nose.tools import eq_, raises, assert_raises  # @UnresolvedImport
-from nose.plugins.skip import SkipTest
 from dateutil.tz import tzoffset
+import pytest
 
 from apscheduler.util import *
 
@@ -40,39 +39,36 @@ def meth():
     pass
 
 
-@raises(ValueError)
 def test_asint_invalid_1():
-    asint('5s')
+    pytest.raises(ValueError, asint, '5s')
 
 
-@raises(ValueError)
 def test_asint_invalid_2():
-    asint('shplse')
+    pytest.raises(ValueError, asint, 'shplse')
 
 
 def test_asint_number():
-    eq_(asint('539'), 539)
+    assert asint('539') == 539
 
 
 def test_asint_none():
-    eq_(asint(None), None)
+    assert asint(None) == None
 
 
 def test_asbool_true():
     for val in (' True', 'true ', 'Yes', ' yes ', '1  '):
-        eq_(asbool(val), True)
+        assert asbool(val) == True
     assert asbool(True) is True
 
 
 def test_asbool_false():
     for val in (' False', 'false ', 'No', ' no ', '0  '):
-        eq_(asbool(val), False)
+        assert asbool(val) == False
     assert asbool(False) is False
 
 
-@raises(ValueError)
 def test_asbool_fail():
-    asbool('yep')
+    pytest.raises(ValueError, asbool, 'yep')
 
 
 def test_convert_datetime_date():
@@ -80,96 +76,95 @@ def test_convert_datetime_date():
     datetimeval = convert_to_datetime(dateval, local_tz, None)
     correct_datetime = datetime(2009, 8, 1)
     assert isinstance(datetimeval, datetime)
-    eq_(datetimeval, correct_datetime.replace(tzinfo=local_tz))
+    assert datetimeval == correct_datetime.replace(tzinfo=local_tz)
 
 
 def test_convert_datetime_passthrough():
     datetimeval = datetime(2009, 8, 1, 5, 6, 12)
     convertedval = convert_to_datetime(datetimeval, local_tz, None)
-    eq_(convertedval, datetimeval.replace(tzinfo=local_tz))
+    assert convertedval == datetimeval.replace(tzinfo=local_tz)
 
 
 def test_convert_datetime_text1():
     convertedval = convert_to_datetime('2009-8-1', local_tz, None)
-    eq_(convertedval, datetime(2009, 8, 1, tzinfo=local_tz))
+    assert convertedval == datetime(2009, 8, 1, tzinfo=local_tz)
 
 
 def test_convert_datetime_text2():
     convertedval = convert_to_datetime('2009-8-1 5:16:12', local_tz, None)
-    eq_(convertedval, datetime(2009, 8, 1, 5, 16, 12, tzinfo=local_tz))
+    assert convertedval == datetime(2009, 8, 1, 5, 16, 12, tzinfo=local_tz)
 
 
 def test_datestring_parse_datetime_micro():
     convertedval = convert_to_datetime('2009-8-1 5:16:12.843821', local_tz, None)
-    eq_(convertedval, datetime(2009, 8, 1, 5, 16, 12, 843821, tzinfo=local_tz))
+    assert convertedval == datetime(2009, 8, 1, 5, 16, 12, 843821, tzinfo=local_tz)
 
 
 def test_existing_tzinfo():
     alter_tz = tzoffset('ALTERNATE', -3600)
     dateval = datetime(2009, 8, 1, tzinfo=alter_tz)
-    eq_(convert_to_datetime(dateval, local_tz, None), dateval)
+    assert convert_to_datetime(dateval, local_tz, None) == dateval
 
 
-@raises(TypeError)
 def test_convert_datetime_invalid():
-    convert_to_datetime(995302092123, local_tz, None)
+    pytest.raises(TypeError, convert_to_datetime, 995302092123, local_tz, None)
 
 
-@raises(ValueError)
 def test_convert_datetime_invalid_str():
-    convert_to_datetime('19700-12-1', local_tz, None)
+    pytest.raises(ValueError, convert_to_datetime, '19700-12-1', local_tz, None)
 
 
 def test_timedelta_seconds():
     delta = timedelta(minutes=2, seconds=30)
     seconds = timedelta_seconds(delta)
-    eq_(seconds, 150)
+    assert seconds == 150
 
 
 def test_time_difference_positive():
     earlier = datetime(2008, 9, 1, second=3)
     later = datetime(2008, 9, 1, second=49)
-    eq_(time_difference(later, earlier), 46)
+    assert time_difference(later, earlier) == 46
 
 
 def test_time_difference_negative():
     earlier = datetime(2009, 4, 7, second=7)
     later = datetime(2009, 4, 7, second=56)
-    eq_(time_difference(earlier, later), -49)
+    assert time_difference(earlier, later) == -49
 
 
 class TestDSTTimeDifference(object):
-    def setup(self):
+    @pytest.fixture(scope='class', autouse=True)
+    def timezone(self, request):
+        def finish():
+            del os.environ['TZ']
+            time.tzset()
+
         if hasattr(time, 'tzset'):
             os.environ['TZ'] = 'Europe/Helsinki'
             time.tzset()
-
-    def teardown(self):
-        if hasattr(time, 'tzset'):
-            del os.environ['TZ']
-            time.tzset()
+            request.addfinalizer(finish)
 
     def test_time_difference_daylight_1(self):
         earlier = datetime(2010, 3, 28, 2)
         later = datetime(2010, 3, 28, 4)
-        eq_(time_difference(later, earlier), 3600)
+        assert time_difference(later, earlier) == 3600
 
     def test_time_difference_daylight_2(self):
         earlier = datetime(2010, 10, 31, 2)
         later = datetime(2010, 10, 31, 5)
-        eq_(time_difference(later, earlier), 14400)
+        assert time_difference(later, earlier) == 14400
 
 
 def test_datetime_ceil_round():
     dateval = datetime(2009, 4, 7, 2, 10, 16, 4000)
     correct_answer = datetime(2009, 4, 7, 2, 10, 17)
-    eq_(datetime_ceil(dateval), correct_answer)
+    assert datetime_ceil(dateval) == correct_answer
 
 
 def test_datetime_ceil_exact():
     dateval = datetime(2009, 4, 7, 2, 10, 16)
     correct_answer = datetime(2009, 4, 7, 2, 10, 16)
-    eq_(datetime_ceil(dateval), correct_answer)
+    assert datetime_ceil(dateval) == correct_answer
 
 
 def test_combine_opts():
@@ -178,40 +173,41 @@ def test_combine_opts():
                    'someprefix.opt3': '789'}
     local_opts = {'opt3': 'abc'}
     combined = combine_opts(global_opts, 'someprefix.', local_opts)
-    eq_(combined, dict(opt1='123', opt3='abc'))
+    assert combined == dict(opt1='123', opt3='abc')
 
 
 def test_callable_name():
-    eq_(get_callable_name(test_callable_name), 'test_callable_name')
-    eq_(get_callable_name(DummyClass.staticmeth), 'staticmeth')
-    eq_(get_callable_name(DummyClass.classmeth), 'DummyClass.classmeth')
-    eq_(get_callable_name(DummyClass.meth), 'meth')
-    eq_(get_callable_name(DummyClass().meth), 'DummyClass.meth')
-    eq_(get_callable_name(DummyClass), 'DummyClass')
-    eq_(get_callable_name(DummyClass()), 'DummyClass')
-    assert_raises(TypeError, get_callable_name, object())
+    assert get_callable_name(test_callable_name) == 'test_callable_name'
+    assert get_callable_name(DummyClass.staticmeth) == 'staticmeth'
+    assert get_callable_name(DummyClass.classmeth) == 'DummyClass.classmeth'
+    assert get_callable_name(DummyClass.meth) == 'meth'
+    assert get_callable_name(DummyClass().meth) == 'DummyClass.meth'
+    assert get_callable_name(DummyClass) == 'DummyClass'
+    assert get_callable_name(DummyClass()) == 'DummyClass'
+    pytest.raises(TypeError, get_callable_name, object())
 
 
 def test_obj_to_ref():
-    assert_raises(ValueError, obj_to_ref, DummyClass.meth)
-    assert_raises(ValueError, obj_to_ref, DummyClass.staticmeth)
-    eq_(obj_to_ref(DummyClass.classmeth), 'testutil:DummyClass.classmeth')
-    eq_(obj_to_ref(shelve.open), 'shelve:open')
+    pytest.raises(ValueError, obj_to_ref, DummyClass.meth)
+    pytest.raises(ValueError, obj_to_ref, DummyClass.staticmeth)
+    assert obj_to_ref(DummyClass.classmeth) == 'tests.test_util:DummyClass.classmeth'
+    assert obj_to_ref(shelve.open) == 'shelve:open'
 
 
 def test_inner_obj_to_ref():
     if sys.version_info < (3, 3):
-        raise SkipTest
-    eq_(obj_to_ref(DummyClass.InnerDummyClass.innerclassmeth), 'testutil:DummyClass.InnerDummyClass.innerclassmeth')
+        pytest.skip('This test requires at least Python 3.3')
+    assert obj_to_ref(DummyClass.InnerDummyClass.innerclassmeth) == \
+        'test_util:DummyClass.InnerDummyClass.innerclassmeth'
 
 
 def test_ref_to_obj():
-    eq_(ref_to_obj('shelve:open'), shelve.open)
-    assert_raises(TypeError, ref_to_obj, object())
-    assert_raises(ValueError, ref_to_obj, 'module')
-    assert_raises(LookupError, ref_to_obj, 'module:blah')
+    assert ref_to_obj('shelve:open') == shelve.open
+    pytest.raises(TypeError, ref_to_obj, object())
+    pytest.raises(ValueError, ref_to_obj, 'module')
+    pytest.raises(LookupError, ref_to_obj, 'module:blah')
 
 
 def test_maybe_ref():
-    eq_(maybe_ref('shelve:open'), shelve.open)
-    eq_(maybe_ref(shelve.open), shelve.open)
+    assert maybe_ref('shelve:open') == shelve.open
+    assert maybe_ref(shelve.open) == shelve.open